refactor: revamp emitters
