--- spec/fixtures/vcr_cassettes/Mihari_Emitters_HTTP/_emit/1_2_1.yml
@@ -1,49 +1,0 @@
----
-http_interactions:
-- request:
-    method: post
-    uri: https://httpbin.org/post
-    body:
-      encoding: UTF-8
-      string: '{"title":"title","description":"description","source":"test","artifacts":["1.1.1.1","github.com"],"tags":[]}'
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-      Content-Type:
-      - application/json
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Sun, 20 Mar 2022 02:49:30 GMT
-      Content-Type:
-      - application/json
-      Content-Length:
-      - '741'
-      Connection:
-      - keep-alive
-      Server:
-      - gunicorn/19.9.0
-      Access-Control-Allow-Origin:
-      - "*"
-      Access-Control-Allow-Credentials:
-      - 'true'
-    body:
-      encoding: UTF-8
-      string: "{\n  \"args\": {}, \n  \"data\": \"{\\\"title\\\":\\\"title\\\",\\\"description\\\":\\\"description\\\",\\\"source\\\":\\\"test\\\",\\\"artifacts\\\":[\\\"1.1.1.1\\\",\\\"github.com\\\"],\\\"tags\\\":[]}\",
-        \n  \"files\": {}, \n  \"form\": {}, \n  \"headers\": {\n    \"Accept\": \"*/*\",
-        \n    \"Accept-Encoding\": \"gzip;q=1.0,deflate;q=0.6,identity;q=0.3\", \n
-        \   \"Content-Length\": \"108\", \n    \"Content-Type\": \"application/json\",
-        \n    \"Host\": \"httpbin.org\", \n    \"User-Agent\": \"Ruby\", \n    \"X-Amzn-Trace-Id\":
-        \"Root=1-62369639-0eed2bbf215fcc143ba78276\"\n  }, \n  \"json\": {\n    \"artifacts\":
-        [\n      \"1.1.1.1\", \n      \"github.com\"\n    ], \n    \"description\":
-        \"description\", \n    \"source\": \"test\", \n    \"tags\": [], \n    \"title\":
-        \"title\"\n  }, \n  \"origin\": \"133.218.173.245\", \n  \"url\": \"https://httpbin.org/post\"\n}\n"
-  recorded_at: Sun, 20 Mar 2022 02:49:30 GMT
-recorded_with: VCR 6.1.0
